Refactor debug_native.c
